chore: Fixed page header testing
