--- app/components/page_header_component.rb
@@ -1,12 +1,0 @@
-# frozen_string_literal: true
-
-# Component for displaying a page header with title, subtitle and buttons
-class PageHeaderComponent < ViewComponent::Base
-  renders_one :icon, Viral::IconComponent
-  renders_one :buttons
-
-  def initialize(title:, subtitle:)
-    @title = title
-    @subtitle = subtitle
-  end
-end
