+++ app/components/viral/page_header_component.html.erb
