+++ app/views/groups/index.html.erb
@@ -1,4 +1,4 @@
-<%= render PageHeaderComponent.new(title: t("groups.show.title"), subtitle: t("groups.show.subtitle")) do |component| %>
+<%= render Viral::PageHeaderComponent.new(title: t("groups.show.title"), subtitle: t("groups.show.subtitle")) do |component| %>
   <%= component.with_icon(name: "squares_2x2", classes: "h-14 w-14 text-primary-700") %>
   <%= component.with_buttons do %>
     <%= link_to t(:"groups.index.create_group_button"),
