+++ app/views/groups/members/index.html.erb
@@ -1,4 +1,4 @@
-<%= render PageHeaderComponent.new(title: t(:'.title'), subtitle: t(:'.subtitle', namespace_type: @namespace.type.downcase, namespace_name: @namespace.name)) do |component| %>
+<%= render Viral::PageHeaderComponent.new(title: t(:'.title'), subtitle: t(:'.subtitle', namespace_type: @namespace.type.downcase, namespace_name: @namespace.name)) do |component| %>
   <%= component.with_icon(name: "users", classes: "h-14 w-14 text-primary-700") %>
   <%= component.with_buttons do %>
     <%= link_to t(:".add"),
