+++ app/views/groups/new.html.erb
@@ -1,4 +1,4 @@
-<%= render PageHeaderComponent.new(title: t(:"groups.create.title"), subtitle: t(:"groups.create.subtitle")) do |component| %>
+<%= render Viral::PageHeaderComponent.new(title: t(:"groups.create.title"), subtitle: t(:"groups.create.subtitle")) do |component| %>
   <%= component.with_icon(name: "squares_plus", classes: "h-14 w-14 text-primary-700") %>
 <% end %>
 <div class="p-4">
