+++ app/views/groups/new_subgroup.html.erb
@@ -1,4 +1,4 @@
-<%= render PageHeaderComponent.new(title: t(:'groups.new_subgroup.title'), subtitle: t(:'groups.new_subgroup.subtitle')) do |component| %>
+<%= render Viral::PageHeaderComponent.new(title: t(:'groups.new_subgroup.title'), subtitle: t(:'groups.new_subgroup.subtitle')) do |component| %>
   <%= component.with_icon(name:"squares_plus", classes: "h-14 w-14 text-primary-700" ) %>
 <% end %>
