+++ app/views/groups/show.html.erb
@@ -1,4 +1,4 @@
-<%= render PageHeaderComponent.new(title: @group.name, subtitle: @group.description) do |component| %>
+<%= render Viral::PageHeaderComponent.new(title: @group.name, subtitle: @group.description) do |component| %>
   <%= component.with_icon(name: "squares_2x2", classes: "h-14 w-14 text-primary-700") %>
   <%= component.with_buttons do %>
     <div class="flex flex-row">
