+++ app/views/projects/index.html.erb
@@ -1,4 +1,4 @@
-<%= render PageHeaderComponent.new(title: t("projects.index.title"), subtitle: t("projects.index.subtitle")) do |component| %>
+<%= render Viral::PageHeaderComponent.new(title: t("projects.index.title"), subtitle: t("projects.index.subtitle")) do |component| %>
   <%= component.with_icon(
     name: "rectangle_stack",
     classes: "h-14 w-14 text-primary-700"
