+++ app/views/projects/new.html.erb
@@ -1,4 +1,4 @@
-<%= render PageHeaderComponent.new(title: t(:".title"), subtitle: t(:".subtitle")) do |component| %>
+<%= render Viral::PageHeaderComponent.new(title: t(:".title"), subtitle: t(:".subtitle")) do |component| %>
   <%= component.with_icon(name: "rectangle_stack", classes: "h-14 w-14 text-primary-700")  %>
 <% end %>
