+++ app/views/projects/samples/index.html.erb
@@ -1,4 +1,4 @@
-<%= render PageHeaderComponent.new(title: t('.title'), subtitle: t(".subtitle")) do |component| %>
+<%= render Viral::PageHeaderComponent.new(title: t('.title'), subtitle: t(".subtitle")) do |component| %>
   <%= component.with_icon(name: "beaker", classes: "h-14 w-14 text-primary-700" ) %>
   <%= component.with_buttons do %>
     <%= link_to t('projects.samples.index.new_button'), new_namespace_project_sample_path, class: "btn btn-primary", 'aria-label': t('.actions.button_add_aria_label') %>
