+++ app/views/projects/samples/new.html.erb
@@ -1,4 +1,4 @@
-<%= render PageHeaderComponent.new(title: t("projects.samples.new.title"), subtitle: t("projects.samples.new.subtitle")) do |component| %>
+<%= render Viral::PageHeaderComponent.new(title: t("projects.samples.new.title"), subtitle: t("projects.samples.new.subtitle")) do |component| %>
   <%= component.with_icon(name: "beaker", classes: "h-14 w-14 text-primary-700") %>
 <% end %>
 <div class="p-4">
