+++ app/views/projects/samples/show.html.erb
@@ -1,4 +1,4 @@
-<%= render PageHeaderComponent.new(title: @sample.name, subtitle: @sample.description) do |component| %>
+<%= render Viral::PageHeaderComponent.new(title: @sample.name, subtitle: @sample.description) do |component| %>
   <%= component.with_icon(name: "beaker", classes: "h-14 w-14 text-primary-700") %>
   <%= component.with_buttons do %>
     <div class="flex flex-row">
