+++ app/views/projects/show.html.erb
@@ -1,4 +1,4 @@
-<%= render PageHeaderComponent.new(title: @project.name, subtitle: @project.description) do |component| %>
+<%= render Viral::PageHeaderComponent.new(title: @project.name, subtitle: @project.description) do |component| %>
   <%= component.with_icon(
     name: "clipboard_document",
     classes: "h-14 w-14 text-primary-700"
